Added rudimentary save system
Intructions: after adding the desired configuration of potentials/particles (not necessarily of both) press the 'Save' button. A 'save.dat' file will be created at the directory where '2dsim.py' is located. If you wish to backup your save rename 'save.dat' to something else (without changing the extension), if you don't do this the save will be overwritten. If you wish to load the configuration simply rename the desired save to 'save.dat' and press the 'Load' button (you should be able to do this with the app running). You still need to press compute after loading the save. A test save file has been included with 3 potentials and 3 particles.
Requires the 'pickle' module, but it should already be installed if the instructions on the 'README.md' file have been followed.